Handle 'credential_id' through 'credential_data' and 'ci_credential_id'
The API property 'credential_id' might contain either a hash with the
user_id/token data or a string with the ID of the CI credential. Thus,
the API server creates an internal record of 'credential_data' to handle
the user_id/token data, generates a name for the CI credential, and adds
it to the 'project_repos' and 'credentials' in the JePL record.